Removed exclusion of .classpath and .project from .gitignore
When importing the project onto a new computer, I was getting several
errors, which I was able to reproduce regularly on two devices. If I
made a new project, deleted the files in it, and then pulled from
GitHub, I received an error, stating that the .project is missing. I
then restored it, and everything appears to be working properly.